+++ src/model/player.hpp
@@ -2,12 +2,12 @@
 
 class Player {
 	private:
-		int pos_x;
-		int pos_y;
+		float pos_x;
+		float pos_y;
 
 	public:
-		int get_pos_x();
-		int get_pos_y();
-		void move(int delta_x, int delta_y);
+		float get_pos_x();
+		float get_pos_y();
+		void move(float delta_x, float delta_y);
 
 };
